Create homework 1.cpp
homework1 (Настя Оперчук)
